# Optimize item Remove operations