… src/Build/Microsoft.Build.csproj
… src/Build/Utilities/FileSpecMatchTester.cs
… src/Shared/Traits.cs