Fixes #6696

### Context

Removing items by referencing other item types or by multiple single-item Remove operations has bad performance characteristics because it leads to N*M comparisons, each involving path normalization of both sides. For projects with as few as small thousands of items, evaluating them becomes a bottleneck with direct impact on build time and VS responsiveness.

### Changes Made

Replaced `ImmutableList<ItemData>` with a new collection type `OrderedItemDataCollection` to be used when representing the intermediate list of items. In addition to being a list, the new collection also lazily creates a dictionary of items keyed by normalized paths, which allows for fast matching.

### Testing

Functional:
- Existing unit tests, one new unit test for a previously uncovered scenario.
- Experimental VS insertion.

Performance:
Used a benchmark project with 2000 files in a subdirectory.

Evaluation time of:
```xml
    <A Include="**/*.*" />
    <B Include="**/*.*" />
    <B Remove="@(A)" />
```
went from 7300 ms down to 180 ms.

Evaluation time of:
```xml
    <A Include="**/*.*" />    
    <A Remove='subdir\file1'/>
    <A Remove='subdir\file2'/>
    ...
    <A Remove='subdir\file2000'/>
```
went from 7900 ms down to 330 ms.

Evaluation time of a Hello World .NET Core project went got slightly faster as well (520 -> 515 ms, note that this was without NGEN so most of it is JITting).

### Notes

I opted for a rather simple implementation of `OrderedItemDataCollection` based on `ImmutableList` and `Dictionary`. It has limitations but seems to work well for the scenarios of interest. It should be possible to implement a completely custom data structure for even better results (requirements: immutable, maintains insertion order for enumeration).