@ladipro 2021-08-16
I've made the following changes:
- No longer maintaining the dictionary as immutable so it is forgotten when the list is added to the cache in `MemoizedOperation`. As @Therzok pointed out earlier, `ImmutableDictionary` has a cost and it turns out that we rarely take advantage of this memoization in real-world projects (when building OC I see <300 cache hits where we further modify the list as opposed to 20k hits with no modification).
- Enabled the new code path only for larger items lists (env variable configurable, >= 100 items by default) because the dictionary set-up work still showed in profiles for small ones.

I have updated the description with fresh measurements. There is a positive impact on evaluation time even if the dictionary is not used, due to other smaller optimizations, primarily caching the normalized value on `ItemData` and refactoring `FileSpecMatchTester` to prevent repeated normalizations.